Feature/initial project hj

See merge request s12-webmobile1-sub1/S12P11E106!36

+++ backend/yourcolors/src/main/java/com/ssafy/yourcolors/domain/info/repository/ProductColorRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductColorRepository extends JpaRepository<ProductColor, ProductColorId> {
-    List<ProductColor> findByProduct_ProductId(Long productId);
+    List<ProductColor> findByProduct_ProductId(int productId);
 }
